@@ -7,25 +7,26 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>3.1.6</version>
+    <version>3.2.0</version>
     <relativePath/> <!-- lookup parent from repository -->
   </parent>
 
   <groupId>com.geneaazul</groupId>
   <artifactId>gedcom-analyzer</artifactId>
-  <version>1.9.0</version>
+  <version>1.9.1</version>
   <name>Gedcom Analyzer</name>
   <description>Gedcom Analyzer by Genea Azul</description>
   <url>https://github.com/genea-azul/gedcom-analyzer</url>
 
   <properties>
-    <java.version>17</java.version>
+    <java.version>21</java.version>
     <google-api-drive.version>v3-rev20231128-2.0.0</google-api-drive.version>
     <gedcom.version>1.14.0</gedcom.version>
     <docker-java.version>3.3.4</docker-java.version>
     <icu.version>74.1</icu.version>
     <pdfbox.version>3.0.1</pdfbox.version>
     <commons-collections4.version>4.4</commons-collections4.version>
+    <commons-compress.version>1.25.0</commons-compress.version>
     <commons-csv.version>1.10.0</commons-csv.version>
     <commons-io.version>2.15.1</commons-io.version>
     <commons-exec.version>1.3</commons-exec.version>
